+++ packages/circular-dependency-scanner/bin/ds.js
@@ -13,11 +13,6 @@ program
   .argument('[path]', 'command execute path. (default: process.cwd())')
   .option('--filter <pattern>', 'glob pattern to filter output circles.')
   .option(
-    '--alias <pairs...>',
-    'path alias, follows `<from>:<to>` convention.',
-    ['@:src'],
-  )
-  .option(
     '--absolute',
     'print absolute path instead. usually use with editor which can quickly jump to the file.',
     false,
@@ -31,11 +26,10 @@ program
   ])
   .option('-t, --throw', "exit with code 1 when cycles're found.", false)
   .action(async (cwd, options) => {
-    const { alias, output, throw: isThrow, ...rest } = options;
+    const { output, throw: isThrow, ...rest } = options;
     const cycles = await circularDepsDetect({
       ...rest,
       cwd,
-      alias: Object.fromEntries(alias.map((v) => v.split(':'))),
     });
 
     if (!cycles.length) return;